+++ internal/generate/generator/doc.go
@@ -1,0 +1,3 @@
+// Package generator generates functions for each arity supported by Imbue's
+// public API.
+package generator